// src/types/callback.d.ts
/**
 * THIS FILE IS GENERATED BY HAND!
 */
import { PirepState } from '../defs'
import { Pirep, Telemetry } from './types'

export interface CallbackHook {
  meta: Meta

  /**
   * This is called when the script is loaded
   *
   */
  setup(): void

  /**
   * This is called every ~500ms
   *
   * @param pirep
   * @param data
   * @param previousData
   */
  run(pirep: Pirep, data: Telemetry, previousData?: Telemetry): void

  /**
   * This is called when the pirep changes state
   *
   * @param pirep
   * @param data
   * @param newPhase
   * @param oldPhase
   */
  phaseChange(
    pirep: Pirep,
    data: Telemetry,
    newPhase: PirepState,
    oldPhase: PirepState,
  ): void
}

/**
 * Meta information about the aircraft config scripts
 */
export interface Meta {
  /**
   * A unique ID for the script. Should be short, no spaces.
   * If you want it truly unique: https://shortunique.id
   */
  id: string

  /**
   * Name of the script
   */
  name: string

  /**
   * Enabled or not
   */
  enabled: boolean

  /**
   * Author of this config file
   */
  author?: string
}